analytics-service/requirements.txt
django==4.2.9
djangorestframework==3.14.0
psycopg2-binary==2.9.9
google-cloud-pubsub==2.19.0
python-dotenv==1.0.0
requests==2.31.0
